Java strings may not be that long
git-svn-id: https://scm.mmbase.org/mmbase/trunk@24470 91430c5f-d668-0410-9e7d-dd663366dcfd

+++ src/org/mmbase/datatypes/StringDataType.java
@@ -23,7 +23,7 @@ import org.mmbase.util.logging.*;
  *
  * @author Pierre van Rooden
  * @author Michiel Meeuwissen
- * @version $Id: StringDataType.java,v 1.42 2007-06-21 07:32:32 pierre Exp $
+ * @version $Id: StringDataType.java,v 1.43 2007-07-07 12:50:28 michiel Exp $
  * @since MMBase-1.8
  */
 public class StringDataType extends ComparableDataType<String> implements LengthDataType<String> {
@@ -34,7 +34,7 @@ public class StringDataType extends ComparableDataType<String> implements Length
     protected PatternRestriction patternRestriction = new PatternRestriction(Pattern.compile("(?s)\\A.*\\z"));
     private boolean isPassword = false;
     protected AbstractLengthDataType.MinRestriction minLengthRestriction = new AbstractLengthDataType.MinRestriction(this, 0);
-    protected AbstractLengthDataType.MaxRestriction maxLengthRestriction = new AbstractLengthDataType.MaxRestriction(this, Long.MAX_VALUE);
+    protected AbstractLengthDataType.MaxRestriction maxLengthRestriction = new AbstractLengthDataType.MaxRestriction(this, Integer.MAX_VALUE);
 
     /**
      * Constructor for string data type.
